+++ app/routes/components/Input/Demo.tsx
@@ -1,5 +1,9 @@
 import { Input } from "src/components/Input";
 
 export function Demo() {
-  return <Input />;
+  return (
+    <Input
+      placeholder="Type something..."
+    />
+  );
 }