+++ app/routes/components/Input/changelog.codegen.ts
@@ -1,3 +1,6 @@
 export const changelog = ''
   + '- 2025-09-11: Initial version.\n'
+  + '- 2025-09-13: Enhance Input component with more styling variables (\n'
+  + '  `--input-font-size`, `--input-line-height`, `--input-caret-color`,\n'
+  + '  `--input-placeholder-color`).\n'
   ;