+++ src/components/Input/CHANGELOG.md
@@ -1,1 +1,4 @@
 - 2025-09-11: Initial version.
+- 2025-09-13: Enhance Input component with more styling variables (
+  `--input-font-size`, `--input-line-height`, `--input-caret-color`,
+  `--input-placeholder-color`).